MOve the compute error array func to live in analysis
