+++ fooof/tests/utils/test_data.py
@@ -19,14 +19,6 @@ def test_trim_spectrum():
     assert np.array_equal(f_out, np.array([2., 3., 4.]))
     assert np.array_equal(p_out, np.array([3., 4., 5.]))
 
-def test_compute_pointwise_error():
-
-    d1 = np.ones(5) * 2
-    d2 = np.ones(5)
-
-    errs = compute_pointwise_error(d1, d2)
-    assert np.array_equal(errs, np.array([1, 1, 1, 1, 1]))
-
 def test_interpolate_spectrum():
 
     freqs, powers = gen_power_spectrum(\
